+++ src/biz/bokhorst/xprivacy/PrivacyManager.java
@@ -544,8 +544,9 @@ public class PrivacyManager {
 			if (methods)
 				for (Hook hook : getHooks(rRestrictionName)) {
 					String settingName = rRestrictionName + "." + hook.getName();
-					String value = getSetting(userId, templateName, settingName, Boolean.toString(parentRestricted)
-							+ (parentAsked ? "+asked" : "+ask"));
+					String value = getSetting(userId, templateName, settingName,
+							Boolean.toString(parentRestricted && !hook.isDangerous())
+									+ (parentAsked ? "+asked" : "+ask"));
 					boolean restricted = value.contains("true");
 					boolean asked = (!ondemand || value.contains("asked"));
 					PRestriction childMerge;